# api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: ESG & Environmental Data Management Platform API
  description: |
    REST API for the ESG Environmental Data Management Platform.
    Manages energy, water, waste, GHG emissions, air emissions, sales, recovery,
    and production data across manufacturing sites.
  version: 2026-02-24
  contact:
    name: ESG Platform Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/2026-02-24
    description: Local development
  - url: http://localhost:3000/api/2026-02-24
    description: Docker (via nginx proxy)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SiteRef:
      type: object
      properties:
        name:
          type: string
        code:
          type: string

    # ── Data Entry Common ──
...
    WaterData:
      type: object
      required: [siteId, month, year]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        freshWaterKl:
          type: number
          minimum: 0
          default: 0
        recycledWaterKl:
          type: number
          minimum: 0
          default: 0
        dischargeKl:
          type: number
          minimum: 0
          default: 0
        totalConsumptionKl:
          type: number
          readOnly: true
        site:
          $ref: '#/components/schemas/SiteRef'
    WasteData:
      type: object
      required: [siteId, month, year, wasteType, disposalMethod, quantityMt]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        wasteType:
          type: string
          enum: [hazardous, non_hazardous]
        disposalMethod:
          type: string
          enum: [recycled, co_processed, landfilled, incinerated, other]
        quantityMt:
          type: number
          minimum: 0
        description:
          type: string
          maxLength: 255
        site:
          $ref: '#/components/schemas/SiteRef'
...
    SalesData:
      type: object
      required: [siteId, month, year, quantityMt]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        quantityMt:
          type: number
          minimum: 0
        revenueLkr:
          type: number
          minimum: 0
        site:
          $ref: '#/components/schemas/SiteRef'
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Duplicate record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []
paths:
...
  /data/water:
    get:
      tags: [Water]
      summary: List water consumption data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of water records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WaterData'
    post:
      tags: [Water]
      summary: Create water entry (auto-calculates total consumption)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaterData'
      responses:
        '201':
          description: Created
        '409':
          $ref: '#/components/responses/Conflict'
  /data/water/{id}:
    put:
      tags: [Water]
      summary: Update water entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Waste ──
  /data/waste:
    get:
      tags: [Waste]
      summary: List waste data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of waste records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WasteData'
    post:
      tags: [Waste]
      summary: Create waste entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WasteData'
      responses:
        '201':
          description: Created
  /data/waste/{id}:
    put:
      tags: [Waste]
      summary: Update waste entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Waste]
      summary: Delete waste entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ETP ──
...
  /data/sales:
    get:
      tags: [Sales]
      summary: List sales data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of sales records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesData'
    post:
      tags: [Sales]
      summary: Create sales entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesData'
      responses:
        '201':
          description: Created
  /data/sales/{id}:
    put:
      tags: [Sales]
      summary: Update sales entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Sales]
      summary: Delete sales entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Recovery ──
...
  /telemetry/error:
    post:
      tags: [Telemetry]
      summary: Report frontend error
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                stack:
                  type: string
                componentStack:
                  type: string
                url:
                  type: string
                userAgent:
                  type: string
                timestamp:
                  type: string
      responses:
        '204':
          description: Error logged
  /telemetry/vitals:
    post:
      tags: [Telemetry]
      summary: Report web-vital performance metric
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  enum: [CLS, FID, FCP, LCP, TTFB, INP]
                value:
                  type: number
                id:
                  type: string
                navigationType:
                  type: string
                rating:
                  type: string
                  enum: [good, needs-improvement, poor]
      responses:
        '204':
          description: Metric logged